+++ src/main/java/org/biiig/dmgm/impl/operators/subgraph_mining/common/GroupByGraphIdListEmbeddings.java
@@ -3,7 +3,7 @@ package org.biiig.dmgm.impl.operators.subgraph_mining.common;
 import de.jesemann.paralleasy.collectors.GroupByKeyListValues;
 
 class GroupByGraphIdListEmbeddings
-  extends GroupByKeyListValues<DFSEmbedding, Integer, DFSEmbedding> {
+  extends GroupByKeyListValues<DFSEmbedding, Long, DFSEmbedding> {
 
   GroupByGraphIdListEmbeddings() {
     super(DFSEmbedding::getGraphId, e -> e);
